CORE-2544, CORE-2545: Move osgi-integration-test to integrationTest
