--- libs/serialization/serialization-amqp/osgi-integration-tests/test-serializable-2/src/main/kotlin/net/corda/bundle2/Document.kt
@@ -1,3 +1,0 @@
-package net.corda.bundle2
-
-data class Document(val content: String)
