--- libs/serialization/serialization-amqp/osgi-integration-tests/test-serializable-4/detekt-baseline.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" ?>
-<SmellBaseline>
-  <ManuallySuppressedIssues></ManuallySuppressedIssues>
-  <CurrentIssues></CurrentIssues>
-</SmellBaseline>
