+++ libs/serialization/serialization-amqp/serializable-cpk-four/detekt-baseline.xml
