+++ libs/serialization/serialization-amqp/serializable-cpk-four/src/main/kotlin/net/corda/bundle4/Transfer.kt
