+++ libs/serialization/serialization-amqp/serializable-cpk-one/build.gradle
