+++ libs/serialization/serialization-amqp/serializable-cpk-one/detekt-baseline.xml
