+++ libs/serialization/serialization-amqp/serializable-cpk-one/src/main/kotlin/net.corda.bundle1/Cash.kt
