+++ libs/serialization/serialization-amqp/serializable-cpk-three/build.gradle
@@ -17,7 +17,7 @@ cordapp {
 }
 
 dependencies {
-    cordapp project(":libs:serialization:serialization-amqp:osgi-integration-tests:test-serializable-1")
+    cordapp project(":libs:serialization:serialization-amqp:serializable-cpk-one")
     cordaProvided platform("net.corda:corda-api:$cordaApiVersion")
     cordaProvided 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
 }
