+++ libs/serialization/serialization-amqp/serializable-cpk-three/detekt-baseline.xml
