+++ libs/serialization/serialization-amqp/serializable-cpk-three/src/main/java/net/corda/bundle3/package-info.java
