+++ libs/serialization/serialization-amqp/serializable-cpk-three/src/main/kotlin/net/corda/bundle3/Obligation.kt
