+++ libs/serialization/serialization-amqp/serializable-cpk-two/build.gradle
