+++ libs/serialization/serialization-amqp/serializable-cpk-two/detekt-baseline.xml
