+++ libs/serialization/serialization-amqp/serializable-cpk-two/src/main/kotlin/net/corda/bundle2/Document.kt
@@ -1,0 +1,3 @@
+package net.corda.bundle2
+
+class Document(val content: String)
