+++ libs/serialization/serialization-amqp/test.bndrun
@@ -6,6 +6,10 @@ java: ${project.javaExecutable}
 -runfw: org.apache.felix.framework
 -resolve.effective: active
 -runtrace: true
+
+# Enable debugging.
+# -runjdb: 5005
+
 -runsystempackages: \
     sun.net.www.protocol.jar,\
     sun.security.x509, \
@@ -13,7 +17,7 @@ java: ${project.javaExecutable}
     javax.xml.stream.events;version=1.0.0,\
     javax.xml.stream.util;version=1.0.0
 -runrequires: \
-    bnd.identity;id='amqp-osgi-tests',\
+    bnd.identity;id='serialization-amqp-tests',\
     bnd.identity;id='junit-jupiter-engine',\
     bnd.identity;id='junit-platform-launcher',\
     bnd.identity;id='slf4j.simple'
@@ -23,7 +27,6 @@ java: ${project.javaExecutable}
     order=sortbynameversion,\
     begin=-1
 -runbundles: \
-	amqp-osgi-tests;version='[5.0.0,5.0.1)',\
 	antlr.osgi;version='[2.7.7,2.7.8)',\
 	bcpkix;version='[1.69.0,1.69.1)',\
 	bcprov;version='[1.69.0,1.69.1)',\
@@ -86,5 +89,6 @@ java: ${project.javaExecutable}
 	org.opentest4j;version='[1.2.0,1.2.1)',\
 	org.osgi.util.function;version='[1.1.0,1.1.1)',\
 	org.osgi.util.promise;version='[1.1.1,1.1.2)',\
+	serialization-amqp-tests;version='[5.0.0,5.0.1)',\
 	slf4j.api;version='[1.7.32,1.7.33)',\
 	slf4j.simple;version='[1.7.32,1.7.33)'
