+++ settings.gradle
@@ -123,11 +123,10 @@ include 'libs:security-manager:osgi-integration-tests:local-permissions-impl'
 include 'libs:serialization:checkpoint-serialization-api'
 include 'libs:serialization:kryo-custom-serializers'
 include 'libs:serialization:serialization-amqp'
-include 'libs:serialization:serialization-amqp:osgi-integration-tests'
-include 'libs:serialization:serialization-amqp:osgi-integration-tests:test-serializable-1'
-include 'libs:serialization:serialization-amqp:osgi-integration-tests:test-serializable-2'
-include 'libs:serialization:serialization-amqp:osgi-integration-tests:test-serializable-3'
-include 'libs:serialization:serialization-amqp:osgi-integration-tests:test-serializable-4'
+include 'libs:serialization:serialization-amqp:serializable-cpk-one'
+include 'libs:serialization:serialization-amqp:serializable-cpk-two'
+include 'libs:serialization:serialization-amqp:serializable-cpk-three'
+include 'libs:serialization:serialization-amqp:serializable-cpk-four'
 include 'libs:serialization:serialization-encoding'
 include 'libs:serialization:serialization-kryo'
 include 'libs:utilities'
